requirements.txt
asgiref==3.7.2
Django==4.2.4
django-appconf==1.0.5
django-environ==0.10.0
django-imagekit==5.0.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
pilkit==3.0
Pillow==10.1.0
psycopg2==2.9.7
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
